third_party/blink/web_tests/external/wpt/css/css-color-adjust/meta/meta-color-scheme-single-value-in-head-expected.txt

This is a testharness.js-based test.
FAIL Computed root color-scheme should be dark assert_equals: expected (string) "dark" but got (undefined) undefined
Harness: the test ran to completion.
